requirements.txt
requests
websocket-client
pandas
numpy
firebase-admin
python-dotenv
schedule